% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schoolgrowth-Package.R
\docType{package}
\name{schoolgrowth-package}
\alias{schoolgrowth}
\alias{schoolgrowth-package}
\title{schoolgrowth: Functions for Improving Accuracy of School-Level Growth Measures Using Empirical Best Linear Prediction}
\description{
Provides tools for smoothing school-level growth measures across years, grades and/or subjects using empirical best linear prediction methods.
}
